More work on IR
With this change, basic generation of IR works for a trivial shader, and there is some basic support for dumping the generated IR in an assembly-like format.

As with the other IR change, the use of the IR is statically disabled for now, so that existing users won't be affected.