[ir] Represent fields more direclty
Previously, a `StructType` was an ordinary instruction that took a variable number of types are operands, representing the types of fields.
This ends up being inconvenient for a few reasons:

- To add decorations to the fields, you'd end up having to decorate the struct type instead (SPIR-V has this problem)

- You need to compute field indices during lowering, when you might prefer to defer that until later

- The get/set field operations now need an index, which needs to be an explicit operand, which means a magic numeric literal floating around to represent the index

The new approach fixes for the first two of these, and at least makes the last one a bit nicer.
A `StructDecl` is now a parent instruction, and its sub-instructions represent the fields of the type - each field is an explicit instruction of type `StructField`.
The operation to extract a field takes a direct reference the struct field, so everything is quite explicit.