Fixup: handle splice of multiple modifiers
I changed the logic so that it might splice a new modifier into the existing linked list (not just at the end), but failed to account for the case where what we are splicing in isn't just a single modifier, but a whole *list* (which occurs when splicing in the shared modifiers themselves).

This change handles that case with a bit of annoying linked-list cleverness.