Fixup for test failure
AppVeyor has a different version of fxc installed by default, and it produces subtly different output for this test case.

It seems like later versions are clever enough to completely eliminate an empty `cbuffer` declaration, but earlier versions aren't.
I'm actually not entirely sure why Slang is successfully eliminating the cbuffer as well, but the output DXBC implies it was not generated.