Fixup: deal with hitting `.obj` size limits for VS
When using the lumped/"unity" build approach for Slang, the resulting `.obj` files run into number-of-sections limits in the VS linker. For now I'm using the `/bigobj` command-line flag to work around this for the `hello` example, just so I can be sure the lumped build still works, but longer term it seems like we need to just drop that approach anyway.

The `render-test` application was switched to link against `slang.dll` since there is no reason to have multiple apps use the lumped approach.