Parser: fix precedence of cast expressions (#203)
We were accidentally parsing this:

    (uint) a / b

as this:

    (uint) ( a / b )

when it should be:

    ((uint) a) / b

This is a bug that seems to have been inherited from a long time ago. It has taken a while to bite anybody because the only class of expressions it would hit are multiplicative ones, and in many cases the difference in the cast order won't be noticed for values in a limited range.