Bug fix for vector initializer lists
When a vector was initialized with an initializer list:

    float4 f = { 0, 1, 2, 3 };

we were following the logic for `struct` types (since `vector<T,N>` is technically a `struct` declaration in our stdlib...), but the type has no field, so we were (silently!) ignoring the actual operands.

I've applied a simple fix where we cast the operands to the element type of the vector, but a more complete fix will be needed sooner or later where we check the operand counts properly, etc.