Move reflection JSON generation into separate text fixture
The ability to dump a JSON listing of reflection info used to be built into the core compiler, but in a somewhat hacky fashion. Now it is instead built as a separate test application that uses the public Slang APIs to do its job, thus serving as an example of what a user could build.