Reflection: allow querying of semantics on varying input/output
This is functionality required to support a Falcor bug fix.

Most of the code to compute the right semantic name/index for a parameter was already present.
This change adds:

- Storage for semantic name/index on every `VarLayout`
  - Note: this is wasteful and should be optimized later

- A public API to query the semantic name/index
  - The contract is that this API returns `NULL` if the parameter had no semantic

- A bit of work in `parameter-binding.cpp` to attach semantics to varying input/output when traversing varying parameters.
  - Note: this is intentionally set up so that it associates semantics even with non-leaf parameters, so that an API user can query the semantic of a `struct` parameter and know that its members will be assigned sequential semantic indices from its starting value.

- Support for dumping this information in reflection tests

One notable thing that I did *not* change here is that the reflection test fixture doesn't report information on the output of an entry point, even though it really should. That should be fixed in a separate change, though, because it would affect many of the expected outputs.